requirements.txt
autopep8==1.7.0
Faker==16.4.0
numpy==1.23.4
pandas==1.5.0
pycodestyle==2.9.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
six==1.16.0
SQLAlchemy==1.4.42
toml==0.10.2
